+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 1.1 (2020-09-05)
+
+* [Add] Added Rule set for automatic Bucket assignments #5
+* [Add] Enabled movement of Buckets to other Bucket Groups
+* [Add] Enabled movement of Bucket Groups #7
+* [Changed] Moved Bucket Movements to Transaction Dialog #1
+* [Changed] Opening a new file resets previous Import selection and settings
+* [Changed] Optimized Y-Axis Ticks for Month Bucket Expenses on Report Page
+
 ### 1.0 (2020-08-10)
 
 * First major release